(No Logic) Move existing premium.component into new premium folder
