+++ apps/web/src/app/billing/individual/individual-billing.module.ts
@@ -6,7 +6,7 @@ import { BillingSharedModule } from "../shared";
 import { BillingHistoryViewComponent } from "./billing-history-view.component";
 import { IndividualBillingRoutingModule } from "./individual-billing-routing.module";
 import { PremiumV2Component } from "./premium/premium-v2.component";
-import { PremiumComponent } from "./premium.component";
+import { PremiumComponent } from "./premium/premium.component";
 import { SubscriptionComponent } from "./subscription.component";
 import { UserSubscriptionComponent } from "./user-subscription.component";
